improve api/impl system
