+++ modules/features/rating/api/build.gradle.kts
@@ -1,0 +1,29 @@
+@file:Suppress("UnusedPrivateMember")
+
+import ru.astrainteractive.gradleplugin.property.extension.ModelPropertyValueExt.requireProjectInfo
+
+plugins {
+    id("com.android.library")
+    kotlin("multiplatform")
+    id("ru.astrainteractive.gradleplugin.java.core")
+    id("ru.astrainteractive.gradleplugin.android.core")
+    alias(libs.plugins.kotlin.serialization)
+}
+
+kotlin {
+    androidTarget()
+    applyDefaultHierarchyTemplate()
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                // Coroutines
+                implementation(libs.kotlin.coroutines.core)
+                implementation(libs.decompose.core)
+                implementation(projects.modules.services.apiEmpireapi)
+            }
+        }
+    }
+}
+android {
+    namespace = "${requireProjectInfo.group}.rating.api"
+}
