+++ modules/features/rating/api/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/rating/user/presentation/RatingUserComponent.kt
