+++ modules/features/rating/api/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/rating/users/di/RatingUsersModule.kt
@@ -1,0 +1,11 @@
+package com.makeevrserg.empireprojekt.mobile.features.rating.users.di
+
+import com.arkivanov.decompose.ComponentContext
+import com.makeevrserg.empireprojekt.mobile.features.rating.users.presentation.RatingUsersComponent
+
+interface RatingUsersModule {
+    fun createRatingUsersComponent(
+        componentContext: ComponentContext,
+        onShowUserRatingsClicked: (userId: Long, userName: String) -> Unit
+    ): RatingUsersComponent
+}
