+++ modules/features/root/ui/build.gradle.kts
@@ -36,15 +36,18 @@ kotlin {
                 implementation(projects.modules.services.core)
                 implementation(projects.modules.features.root.api)
                 implementation(projects.modules.features.root.impl)
+                implementation(projects.modules.features.rating.api)
                 implementation(projects.modules.features.rating.impl)
                 implementation(projects.modules.features.rating.ui)
                 implementation(projects.modules.features.splash.impl)
                 implementation(projects.modules.features.splash.ui)
+                implementation(projects.modules.features.status.api)
                 implementation(projects.modules.features.status.impl)
                 implementation(projects.modules.features.status.ui)
                 implementation(projects.modules.features.theme.api)
                 implementation(projects.modules.features.theme.impl)
                 implementation(projects.modules.features.theme.ui)
+                implementation(projects.modules.features.towns.api)
                 implementation(projects.modules.features.towns.impl)
                 implementation(projects.modules.features.towns.ui)
                 implementation(projects.modules.features.info.ui)
