+++ modules/features/status/api/build.gradle.kts
@@ -1,0 +1,28 @@
+@file:Suppress("UnusedPrivateMember")
+
+import ru.astrainteractive.gradleplugin.property.extension.ModelPropertyValueExt.requireProjectInfo
+
+plugins {
+    id("com.android.library")
+    kotlin("multiplatform")
+    id("ru.astrainteractive.gradleplugin.java.core")
+    id("ru.astrainteractive.gradleplugin.android.core")
+    alias(libs.plugins.kotlin.serialization)
+}
+
+kotlin {
+    androidTarget()
+    applyDefaultHierarchyTemplate()
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                // Coroutines
+                implementation(libs.kotlin.coroutines.core)
+                implementation(libs.moko.resources.core)
+            }
+        }
+    }
+}
+android {
+    namespace = "${requireProjectInfo.group}.status.api"
+}
