+++ modules/features/status/api/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/status/di/StatusModule.kt
@@ -1,0 +1,8 @@
+package com.makeevrserg.empireprojekt.mobile.features.status.di
+
+import com.makeevrserg.empireprojekt.mobile.features.status.root.presentation.RootStatusComponent
+
+interface StatusModule {
+
+    val rootStatusComponent: RootStatusComponent
+}
