+++ modules/features/status/impl/build.gradle.kts
@@ -35,6 +35,7 @@ kotlin {
                 implementation(libs.ktor.client.core)
                 // Local
                 implementation(projects.modules.services.core)
+                implementation(projects.modules.features.status.api)
             }
         }
