+++ modules/features/status/impl/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/status/di/StatusModuleImpl.kt
@@ -1,0 +1,22 @@
+package com.makeevrserg.empireprojekt.mobile.features.status.di
+
+import com.arkivanov.mvikotlin.main.store.DefaultStoreFactory
+import com.makeevrserg.empireprojekt.mobile.features.status.root.presentation.DefaultRootStatusComponent
+import com.makeevrserg.empireprojekt.mobile.features.status.root.presentation.RootStatusComponent
+import com.makeevrserg.empireprojekt.mobile.services.core.di.CoreModule
+
+class StatusModuleImpl(
+    private val coreModule: CoreModule
+) : StatusModule {
+
+    override val rootStatusComponent: RootStatusComponent by lazy {
+        val dependencies = RootStatusComponentDependencies.Default(
+            dispatchers = coreModule.dispatchers,
+            storeFactory = DefaultStoreFactory(),
+            httpClient = coreModule.httpClient
+        )
+        DefaultRootStatusComponent(
+            dependencies = dependencies
+        )
+    }
+}
