+++ modules/features/theme/api/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/theme/di/ThemeSwitcherModule.kt
@@ -1,0 +1,8 @@
+package com.makeevrserg.empireprojekt.mobile.features.theme.di
+
+import com.makeevrserg.empireprojekt.mobile.features.theme.presentation.ThemeSwitcherComponent
+
+interface ThemeSwitcherModule {
+
+    val themeSwitcherComponent: ThemeSwitcherComponent
+}
