+++ modules/features/theme/impl/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/theme/di/ThemeSwitcherModuleImpl.kt
@@ -1,0 +1,16 @@
+package com.makeevrserg.empireprojekt.mobile.features.theme.di
+
+import com.makeevrserg.empireprojekt.mobile.features.theme.data.ThemeSwitcherRepositoryImpl
+import com.makeevrserg.empireprojekt.mobile.features.theme.presentation.DefaultThemeSwitcherComponent
+import com.makeevrserg.empireprojekt.mobile.features.theme.presentation.ThemeSwitcherComponent
+import com.makeevrserg.empireprojekt.mobile.services.core.di.CoreModule
+
+class ThemeSwitcherModuleImpl(coreModule: CoreModule) : ThemeSwitcherModule {
+    override val themeSwitcherComponent: ThemeSwitcherComponent by lazy {
+        DefaultThemeSwitcherComponent(
+            themeSwitcherRepository = ThemeSwitcherRepositoryImpl(
+                settings = coreModule.settings
+            )
+        )
+    }
+}
