+++ modules/features/towns/api/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/features/towny/towns/di/TownsModule.kt
@@ -1,0 +1,8 @@
+package com.makeevrserg.empireprojekt.mobile.features.towny.towns.di
+
+import com.arkivanov.decompose.ComponentContext
+import com.makeevrserg.empireprojekt.mobile.features.towny.towns.presentation.TownsComponent
+
+interface TownsModule {
+    fun createTownsComponent(componentContext: ComponentContext): TownsComponent
+}
