+++ modules/features/towns/impl/build.gradle.kts
@@ -41,6 +41,7 @@ kotlin {
                 implementation(projects.modules.services.coreResources)
                 implementation(projects.modules.services.core)
                 implementation(projects.modules.services.apiEmpireapi)
+                implementation(projects.modules.features.towns.api)
             }
         }
     }
