+++ modules/features/towns/ui/build.gradle.kts
@@ -37,6 +37,7 @@ kotlin {
                 implementation(projects.modules.services.coreUi)
                 implementation(projects.modules.services.core)
                 implementation(projects.modules.services.apiEmpireapi)
+                implementation(projects.modules.features.towns.api)
                 implementation(projects.modules.features.towns.impl)
             }
         }
