+++ modules/services/core-ui/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/core/ui/DecomposeScreen.kt
@@ -1,0 +1,9 @@
+package com.makeevrserg.empireprojekt.mobile.core.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+fun interface DecomposeScreen<T : Any> {
+    @Composable
+    fun Render(modifier: Modifier, child: T)
+}
