+++ settings.gradle.kts
@@ -44,10 +44,13 @@ include(":modules:features:splash:ui")
 include(":modules:features:theme:api")
 include(":modules:features:theme:impl")
 include(":modules:features:theme:ui")
+include(":modules:features:status:api")
 include(":modules:features:status:impl")
 include(":modules:features:status:ui")
+include(":modules:features:rating:api")
 include(":modules:features:rating:impl")
 include(":modules:features:rating:ui")
+include(":modules:features:towns:api")
 include(":modules:features:towns:impl")
 include(":modules:features:towns:ui")
 include(":modules:features:info:ui")
